+++ contracts/modules/andromeda-rates/src/testing/mock_querier.rs
@@ -49,6 +49,7 @@ pub fn mock_dependencies_custom(
     deps
 }
 
+#[allow(dead_code)]
 pub struct WasmMockQuerier {
     pub base: MockQuerier,
     pub contract_address: String,