+++ packages/andromeda-data-storage/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod boolean;
+pub mod form;
 pub mod primitive;
 pub mod string_storage;