+++ packages/andromeda-non-fungible-tokens/src/crowdfund.rs
@@ -118,16 +118,18 @@ pub enum CampaignStage {
     SUCCESS,
     /// Stage when campaign failed to meet the target cap before expiration
     FAILED,
+    /// Stage when campaign is discarded
+    DISCARDED,
 }
 
-impl ToString for CampaignStage {
-    #[inline]
-    fn to_string(&self) -> String {
+impl std::fmt::Display for CampaignStage {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
-            Self::READY => "READY".to_string(),
-            Self::ONGOING => "ONGOING".to_string(),
-            Self::SUCCESS => "SUCCESS".to_string(),
-            Self::FAILED => "FAILED".to_string(),
+            Self::READY => write!(f, "READY"),
+            Self::ONGOING => write!(f, "ONGOING"),
+            Self::SUCCESS => write!(f, "SUCCESS"),
+            Self::FAILED => write!(f, "FAILED"),
+            Self::DISCARDED => write!(f, "DISCARDED"),
         }
     }
 }